backend/requirements.txt
uvicorn
fastapi==0.128.0
uvicorn==0.40.0
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
python-jose==3.5.0
passlib==1.7.4
bcrypt==5.0.0
email-validator==2.3.0
pydantic==2.12.5
python-multipart==0.0.22
boto3==1.28.46
mangum==0.1.0